samples/damon/prcl: fix type of max_age_region
The type of ->max_age_region of struct damos_access_pattern is unsigned
int, but the sample program is setting it as the max value of unsigned
long type.  Fix it to use the max value of unsigned int type.

Fixes: 6a0ed0a7aac3 ("samples/damon/prcl: implement schemes setup")
Reported-by: kernel test robot <lkp@intel.com>
Closes: https://lore.kernel.org/oe-kbuild-all/202411231420.di7hRYuZ-lkp@intel.com/

+++ samples/damon/prcl.c
@@ -83,7 +83,7 @@ static int damon_sample_prcl_start(void)
 			.min_nr_accesses = 0,
 			.max_nr_accesses = 0,
 			.min_age_region = 50,
-			.max_age_region = ULONG_MAX},
+			.max_age_region = UINT_MAX},
 			DAMOS_PAGEOUT,
 			0,
 			&(struct damos_quota){},
